@@ -1,11 +1,9 @@
-using BotLibraryV2;
-using Microsoft.VisualStudio.TestTools.UnitTesting;
-using Microsoft.Bot.Builder;
 using System;
-using System.Resources;
-using System.Threading;
 using System.Globalization;
+using System.Resources;
 using System.Reflection;
+
+using Microsoft.VisualStudio.TestTools.UnitTesting;
 
 namespace BotLibraryTest
 {
@@ -18,14 +16,18 @@ namespace BotLibraryTest
     public class CalenderTests
     {
         private static ResourceManager rm = new ResourceManager("PlanB.Butler.Library.Test.DictionaryTest.resourceTest", Assembly.GetExecutingAssembly());
-              
+        private static string hello = rm.GetString("hello");
+      
         /// <summary>
         /// Called when defined workdays is wrong.
         /// </summary>
         [TestMethod]
         public void OnlyDefinedWorkdaysFail()
         {
-           
+            var hello1 = hello;
+
+            
+
         }
     }
 }
