@@ -87,13 +87,16 @@ namespace PlanB.Butler.Services
                 if (blob != null)
                 {
                     blob.Properties.ContentType = "application/json";
-                    blob.Metadata.Add(MetaDate, date);
+                    var metaDate = mealModel.Date.ToString("yyyyMMdd", CultureInfo.InvariantCulture);
+                    blob.Metadata.Add(MetaDate, metaDate);
                     blob.Metadata.Add(MetaRestaurant, mealModel.Restaurant);
-                    blob.Metadata.Add(Constants.ButlerCorrelationTraceName, correlationId.ToString().Replace("-", string.Empty));
+
+                    //  blob.Metadata.Add(Constants.ButlerCorrelationTraceName, correlationId.ToString().Replace("-", string.Empty));
                     var meal = JsonConvert.SerializeObject(mealModel);
                     trace.Add("meal", meal);
 
                     Task task = blob.UploadTextAsync(requestBody);
+                    task.Wait();
                 }
 
                 actionResult = new OkResult();
